+++ Express/projects/small_projects/crud/src/products/controller.js
@@ -1,67 +1,79 @@
-const Product = require("./model");
+// service.js
+const service = require("./service");
 
-// Create a single product
-async function createProduct(data) {
-  const { name, image, description, price, stock } = data;
-  if (!name || !image || !description || price == null || stock == null) {
-    throw new Error("All fields are required");
+// POST - Create one or more products
+const createProducts = async (req, res) => {
+  try {
+    const body = req.body;
+    const result = Array.isArray(body)
+      ? await service.bulkCreateProducts(body)
+      : await service.createProduct(body);
+    res.status(201).json(result);
+  } catch (err) {
+    res.status(400).json({ error: err.message });
   }
-
-  return await Product.create({ name, image, description, price, stock });
-}
-
-// Bulk create products
-async function bulkCreateProducts(list) {
-  return await Product.bulkCreate(list);
-}
-
-// Get all products
-async function getAllProducts() {
-  return await Product.findAll();
-}
-
-// Get product by ID
-async function getProductById(id) {
-  const product = await Product.findByPk(id);
-  if (!product) throw new Error("Product not found");
-  return product;
-}
-
-// Full update (PUT)
-async function updateProductFull(id, data) {
-  const product = await Product.findByPk(id);
-  if (!product) throw new Error("Product not found");
-
-  const { name, image, description, price, stock } = data;
-  if (!name || !image || !description || price == null || stock == null) {
-    throw new Error("All fields are required for PUT");
-  }
-
-  return await product.update({ name, image, description, price, stock });
-}
-
-// Partial update (PATCH)
-async function updateProductPartial(id, data) {
-  const product = await Product.findByPk(id);
-  if (!product) throw new Error("Product not found");
-
-  return await product.update(data);
-}
-
-// Delete product
-async function deleteProduct(id) {
-  const product = await Product.findByPk(id);
-  if (!product) throw new Error("Product not found");
-
-  await product.destroy();
-}
-
-module.exports = {
-  createProduct,
-  bulkCreateProducts,
-  getAllProducts,
-  getProductById,
-  updateProductFull,
-  updateProductPartial,
-  deleteProduct,
 };
+
+// GET - All products
+const getProducts = async (_req, res) => {
+  try {
+    const products = await service.getAllProducts();
+    res.status(200).json(products);
+  } catch (err) {
+    res.status(500).json({ error: err.message });
+  }
+};
+
+// GET - Product by ID
+const getProductsById = async (req, res) => {
+  try {
+    const product = await service.getProductPartial(req.params.id);
+    res.status(200).json(product);
+  } catch (err) {
+    res.status(404).json({ error: err.message });
+  }
+};
+
+// PUT - Full update
+const updateProducts = async (req, res) => {
+  try {
+    const product = await service.updateProductFull(req.params.id, req.body);
+    res.status(200).json(product);
+  } catch (err) {
+    const status = err.message === "Product not found" ? 404 : 400;
+    res.status(status).json({ error: err.message });
+  }
+};
+
+// PATCH - Partial update
+const updateProductsById = async (req, res) => {
+  try {
+    const product = await service.updateProductPartial(req.params.id, req.body);
+    res.status(200).json(product);
+  } catch (err) {
+    const status = err.message === "Product not found" ? 404 : 400;
+    res.status(status).json({ error: err.message });
+  }
+};
+
+// DELETE - Remove product
+const deleteProducts = async (req, res) => {
+  try {
+    await service.deleteProduct(req.params.id);
+    res.status(204).send();
+  } catch (err) {
+    const status = err.message === "Product not found" ? 404 : 500;
+    res.status(status).json({ error: err.message });
+  }
+};
+
+// Exporting all the controller functions
+module.exports = {
+  createProducts,
+  getProducts,
+  getProductsById,
+  updateProducts,
+  updateProductsById,
+  deleteProducts
+};
+
